Crate aster_network

Re-exports

pub use dma_pool::DmaSegment;

Modules

dma_pool

Structs

EthernetAddr

RxBuffer

TxBuffer

Enums

NetError

Constants

TX_BUFFER_LEN

Statics

RX_BUFFER_POOL

Traits

AnyNetworkDevice

Functions

all_devices

get_device

raise_receive_softirq

Raises softirq for handling reception events

raise_send_softirq

Raises softirq for handling transmission events

register_device

register_recv_callback

Registers callback which will be called when receiving message.

register_send_callback

Registers a callback that will be invoked when the device has completed sending a packet.

Trait Aliases

NetDeviceCallback