aster_network

Function register_recv_callback

pub fn register_recv_callback(name: &str, callback: impl NetDeviceCallback)

Registers callback which will be called when receiving message.

Since the callback will be called in softirq context, the callback function should not sleep.